nexsvg/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.0
starvector==1.0.0
python-jose==3.3.0 